add interface file to the load package
- replace parameters in window methods from interface type to Modal and Page types
